+++ src/modules/bookmarks/middleware/list.js
@@ -1,8 +1,9 @@
 /**
- * @module wallaby/modules/auth/middleware/list
+ * @module wallaby/modules/bookmarks/middleware/list
  */
 
 import { userHistory, userBookmarks } from '../data-keys'
+import { toAppModel } from '../data-utils'
 
 /**
  * Initialize middleware to list Bookmarks.
@@ -37,13 +38,7 @@ function initList(redis, defaultLimit = 10, defaultOffset = 0) {
 
     if (bookmarkIds && bookmarkIds.length) {
       const bookmarks = await redis.hmgetAsync(userBookmarks(user), bookmarkIds)
-      const result = bookmarks
-        .map(bookmarkStr => JSON.parse(bookmarkStr))
-        .map(bookmark => Object.assign(bookmark, {
-          // database stores index, but expose actual values to client
-          page: bookmark.page + 1,
-          panel: bookmark.panel + 1
-        }))
+      const result = toAppModel(bookmarks)
 
       ctx.body = {
         status: 'ok',